requirements.txt
aiodns==1.1.1
aiohttp==3.3.0
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.1.0
cchardet==2.1.1
ccxt==1.14.139
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
idna==2.6
idna-ssl==1.0.1
multidict==4.3.1
pycares==2.3.0
pycparser==2.18
pyOpenSSL==18.0.0
requests==2.18.4
six==1.11.0
urllib3==1.22
yarl==1.1.0
